schnauzer: move legacy api behind v1 namespace
